anv: Fix a harmless overflow warning
anv_pipeline_binding::index is a uint8_t, but some code assigned to it
UINT16_MAX.

Reviewd-by: Jason Ekstrand <jason@jlekstrand.net>

+++ src/intel/vulkan/anv_pipeline.c
@@ -664,7 +664,7 @@ anv_pipeline_compile_fs(struct anv_pipeline *pipeline,
          rt_bindings[0] = (struct anv_pipeline_binding) {
             .set = ANV_DESCRIPTOR_SET_COLOR_ATTACHMENTS,
             .binding = 0,
-            .index = UINT16_MAX,
+            .index = UINT8_MAX,
          };
          num_rts = 1;
       }
